Update and rename to world_cup.py
I hope you don't mind me initiating a pull request. Adding the proper extension to code files (.py for Python files) will properly tag them in GItHub and will help with other things such as integrating with Travis CI as a GitHub Action (or another CI/CD process) if you choose to do so at some point. Also, it is standard in Python to not capitalize function names. I ran it through PyCharm and started auto formatting a couple things at one point. And then that last part, "if __name__ = '__main__':" is something my Python professor taught me. It prepares for scaling up, and creating modules that might call other modules and that is the way it knows to not look or start with main() if that is the case.